+++ src/sw.js
@@ -60,6 +60,10 @@ self.addEventListener( `activate`, event => {
     event.waitUntil( self.clients.claim() )
 } )
 
+self.addEventListener( `message`, event => {
+    if( event.data?.type === `SKIP_WAITING` ) self.skipWaiting()
+} )
+
 self.addEventListener( `fetch`, event => {
 
     const { request } = event